Factor toml formatting logic out into the serialization module
