+++ src/serialization.rs
@@ -109,3 +109,53 @@ pub mod dependency_criteria {
         Ok(m.into_iter().map(|(k, Wrapper(v))| (k, v)).collect())
     }
 }
+
+/// tables with these names will be rendered by TomlFormatter as inline tables,
+/// rather than dotted tables.
+fn always_render_inline(key: &str) -> bool {
+    key == "dependency-criteria"
+}
+
+/// Serialize the given data structure as a formatted `toml_edit::Document`.
+///
+/// The returned document can be converted to a string or otherwise written out
+/// using it's `Display` implementation.
+///
+/// Can fail if `T`'s implementation of `Serialize` fails.
+pub fn to_formatted_toml<T>(val: T) -> Result<toml_edit::Document, toml_edit::ser::Error>
+where
+    T: Serialize,
+{
+    use toml_edit::visit_mut::VisitMut;
+
+    struct TomlFormatter;
+    impl VisitMut for TomlFormatter {
+        fn visit_table_mut(&mut self, node: &mut toml_edit::Table) {
+            // Hide unnecessary implicit table headers for tables containing
+            // only other tables. We don't do this for empty tables as otherwise
+            // they could be hidden.
+            if !node.is_empty() {
+                node.set_implicit(true);
+            }
+            for (k, v) in node.iter_mut() {
+                if !always_render_inline(&k) {
+                    // Try to convert the value into either a table or an array of
+                    // tables if it is currently an inline table or inline array of
+                    // tables.
+                    *v = std::mem::take(v)
+                        .into_table()
+                        .map(toml_edit::Item::Table)
+                        .unwrap_or_else(|i| i)
+                        .into_array_of_tables()
+                        .map(toml_edit::Item::ArrayOfTables)
+                        .unwrap_or_else(|i| i);
+                }
+                self.visit_item_mut(v);
+            }
+        }
+    }
+
+    let mut toml_document = toml_edit::ser::to_document(&val)?;
+    TomlFormatter.visit_document_mut(&mut toml_document);
+    Ok(toml_document)
+}
